+++ packages/back-end/types/visual-changeset.d.ts
@@ -30,10 +30,10 @@ export interface VisualChangesetURLPattern {
 export interface VisualChangesetInterface {
   id: string;
   organization: string;
-  urlPatterns: VisualChangesetURLPattern[]; // single pattern per URL Redirect Visual Changeset
+  urlPatterns: VisualChangesetURLPattern[]; // single URLPattern per URL Redirect Visual Changeset
   editorUrl: string;
   experiment: string;
   visualChanges: VisualChange[];
-  urlRedirects: URLRedirect[];
+  urlRedirects?: URLRedirect[];
   persistQueryString?: boolean;
 }